+++ docs/source/changelog.rst
@@ -11,6 +11,7 @@ Version (dev)
 
 * Fix unknown bits sometimes not round tripping through enum bitfields.
 * Fix incorrect parsing of the deprecated ``bbox()`` FGD helper.
+* Add support for saving/loading VTF hotspot data.
 
 -------------
 Version 2.6.1